+++ frontend/src/components/context/CameraSocketContext.tsx
