--- frontend/src/locales/en.json
@@ -1,4 +1,0 @@
-{
-    "login": "Login",
-    "register": "Register"
-}
