--- frontend/src/locales/id.json
@@ -1,14 +1,0 @@
-{
-  "welcome": {
-    "short": "Selamat Datang di Bye Cheating!"
-  },
-  "login": "Masuk",
-  "to-login-suggestion": "Sudah punya akun, Masuk",
-  "register": "Daftar",
-  "to-register-suggestion": "Tidak punya akun, Daftar!",
-  "textfield": {
-    "email": "Email",
-    "password": "Password",
-    "name": "Nama"
-  }
-}
