+++ frontend/src/routes/app.tsx
@@ -1,5 +1,5 @@
 import { ErrorScreen } from "@/components/ui/ErrorScreen";
-import { AsideContextProvider } from "@/context/AsideContext";
+import { AsideContextProvider } from "@/components/context/AsideContext";
 import MainLayout from "@/layout/MainLayout";
 import { Button, Center } from "@mantine/core";
 import { createFileRoute, useRouter } from "@tanstack/react-router";
